- restructure data for optimization (continuing)   - time vector
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@19866 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/optimization/localFunction.h
@@ -73,7 +73,7 @@ int diff_symColoredObject(IPOPT_DATA_ *iData, double *gradF, int this_it);
 
 /*model*/
 int functionODE_(double * x, double *u, double t, double * dotx, IPOPT_DATA_ *iData);
-int refreshSimData(double *x, double *u, double t, IPOPT_DATA_ *iData);
+int refreshSimData(double *x, double *u, long double t, IPOPT_DATA_ *iData);
 
 /*allocate*/
 int loadDAEmodel(DATA* data, IPOPT_DATA_ *iData);
